+++ odoo/local-src/specific_stock/__manifest__.py
@@ -34,10 +34,13 @@
         'views/stock_production_lot.xml',
         'views/stock_config_settings.xml',
         'views/stock_quant_package.xml',
-        'wizards/stock_receive_lot.xml',
 
         # Wizards
+        'wizards/stock_receive_lot.xml',
         'wizards/stock_backorder_choice.xml',
+        'wizards/put_in_pack_helper.xml',
+
+        'views/stock_picking.xml',
 
         # Security
         'security/ir.model.access.csv',